Fixed small bug in TextGenerator.
In the case that all the words have been previously picked
for a given min/max range, the text generator would fail
with an exception. It has been corrected to detect this
and randomly pick one of the previously used words.

This is a problem that exhibits itself when the dictionary
is relatively small.