Fixed another small bug in the ReadTrainer.
With small dictionaries, there can be an issue where the min and max
are the same and no word(s) match. In this edge case, the max should
be pushed up until one or more words are returned.

For example, if 2-letter US states compose the active dictionary
and the min/max is set to 3. The max should be bumped to 5 since
that is the next increment.