Fixed a couple of bugs when resetting local storage.
I forced the settings page to reload since the controls were not
automatically picking up the updates. I suppose it could be done
more smartly. How often do people do this?

Also made sure that the active dictionary size is no bigger than
the total size of the dictionary on reset. Since not all dictionary
entries are active by default, it is possible for this to happen.