Configurable Dictionary (#21)
* Allow the user to limit words in the dictionary.

Some users may wish to try and learn the sound patterns
of the most common words, whereas other users will want
to focus on recognizing a large number of words. This
feature caters to both sets of users.

* Tagged dictionary with entry type.

This will allow customization on which part(s) of
the dictionary should be used.

* Added extra Q-Codes and Common US First Names

* Intermediate work

Working on allowing the user to select a subset of the dictionary.

* Refactoring to encapsulate dictionary related data

* Dictionary subselections partially works.

Still need to store settings associated with subselections
and ensure that the dictionary is correctly initialized.

* Added storage of dictionary subsets.

There are known bugs that need to be fixed.

* Added a handful of common'ish abbreviations.

* Added US State Abbreviations

* Added countries to dictionary

* Fixed TextGenerator bug.

When the active dictionary contains fewer words, there is a possibility
that a single word will not fit within the min/max range. So I have
lossened the selection of words while selecting multiple words.

* Fixed small bug in TextGenerator.

In the case that all the words have been previously picked
for a given min/max range, the text generator would fail
with an exception. It has been corrected to detect this
and randomly pick one of the previously used words.

This is a problem that exhibits itself when the dictionary
is relatively small.

* Fixed bug in ReadTrainer with same word selected.

The 'state' is associated with the current word. When the same word
is selected, it needs to be different so an extra space is added.

* Direct git to ignore WebStorm configuration

* Fixed another small bug in the ReadTrainer.

With small dictionaries, there can be an issue where the min and max
are the same and no word(s) match. In this edge case, the max should
be pushed up until one or more words are returned.

For example, if 2-letter US states compose the active dictionary
and the min/max is set to 3. The max should be bumped to 5 since
that is the next increment.

* Trivial. Removed commented out line of code.

* Fixed a couple of bugs when resetting local storage.

I forced the settings page to reload since the controls were not
automatically picking up the updates. I suppose it could be done
more smartly. How often do people do this?

Also made sure that the active dictionary size is no bigger than
the total size of the dictionary on reset. Since not all dictionary
entries are active by default, it is possible for this to happen.

* This corrects all known issues.

Refactoring needs to take place to reduce some of the duplication
and tidy up the code.

* Tweaked display of dictionary options on settings page

* First iteration of refactoring to remove duplication.

* Additional refactoring and cleanup

* Setting the min active dictionary size to 10.

This is a more sane value.

* Corrected small bug with the active dictionary size not being saved.

* Update src/ReadTrainer/ReadTrainer.js

Co-Authored-By: zoglmannk <zoglmannk@yahoo.com>

* Minor. Corrected mispellings in dictionary.

Also added TODO.

* Corrected a few other mispelled country names.

* Added comment about potential performance problem and the fix.

* Requested minor cleanup.

Changed setTypes and setType in MorseStore to always
setup the active dictionary so that it isn't
duplicated else where.

* Minor. Changed variable name for readability.