Farnsworth UI slider enhancement
It seems more intuitive to have the maximum value
of the Effective Speed slider updated as the
Character Speed Slider is changed.

This also corrects a UI edge case where the effect
speed will be reset to the default (30wpm)
overriding the previous user value. This can occur
when the slider value has been artificially trimmed.